// src/interfaces/inventoryHistory.interface.ts
export interface IInventoryHistory {
  myKey?: number;
  Item_Number: number;
  Week01?: number;
  Week02?: number;
  Week03?: number;
  Week04?: number;
  Week05?: number;
  Week06?: number;
  Week07?: number;
  Week08?: number;
  Week09?: number;
  Week10?: number;
  Week11?: number;
  Week12?: number;
  Units_Week01?: number;
  Units_Week02?: number;
  Units_Week03?: number;
  Units_Week04?: number;
  Units_Week05?: number;
  Units_Week06?: number;
  Units_Week07?: number;
  Units_Week08?: number;
  Units_Week09?: number;
  Units_Week10?: number;
  Units_Week11?: number;
  Units_Week12?: number;
  Lost01?: number;
  Lost02?: number;
  Lost03?: number;
  Lost04?: number;
  Lost05?: number;
  Lost06?: number;
  Lost07?: number;
  Lost08?: number;
  Lost09?: number;
  Lost10?: number;
  Lost11?: number;
  Lost12?: number;
  Units_Lost01?: number;
  Units_Lost02?: number;
  Units_Lost03?: number;
  Units_Lost04?: number;
  Units_Lost05?: number;
  Units_Lost06?: number;
  Units_Lost07?: number;
  Units_Lost08?: number;
  Units_Lost09?: number;
  Units_Lost10?: number;
  Units_Lost11?: number;
  Units_Lost12?: number;
  Date_LastRecd?: Date;
  Date_LastSold?: Date;
  LastPurchaseVendor?: number;
  LastPurchaseCost?: number;
}